`direct predict` option for inference and minor fixes (#151)
* Replace predict script with `direct predict`
* Check if sewar is installed when using calgary_campinas_vif metric
