+++ projects/calgary_campinas/predict_test.py
@@ -1,5 +1,6 @@
 # coding=utf-8
 # Copyright (c) DIRECT Contributors
+
 import functools
 import logging
 import os
@@ -15,8 +16,6 @@ from direct.data.mri_transforms import Compose
 from direct.environment import Args
 from direct.inference import build_inference_transforms, setup_inference_save_to_h5
 from direct.utils import set_all_seeds
-
-from .utils import volume_post_processing_func
 
 logger = logging.getLogger(__name__)
 
@@ -150,7 +149,6 @@ if __name__ == "__main__":
         args.device,
         args.num_workers,
         args.machine_rank,
-        volume_post_processing_func,
         args.mixed_precision,
         args.debug,
     )
